pyproject.toml
[project]
name = "radarz"
version = "0.1.0"
description = "AI-powered GitHub project recommendation system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.5",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
